requirements.txt
tensorflow
numpy
opencv-python
pyttsx3
pillow
flask
